seperate security.py into multiple different files
